Remove: react-native-svg-charts
