+++ apps/native/components/custom/view/item-details/&pager/&wrapper/index.tsx
@@ -7,8 +7,8 @@ import { createStyleSheet, useStyles } from 'react-native-unistyles';
 import { usePagerViewNavigation } from '@/hooks/usePagerViewNavigation';
 
 import ItemDetailsPagerNavView from '../&nav';
+import ItemDetailsPagerChartPageView from '../&page/chart';
 import ItemDetailsPagerImagePageView from '../&page/image';
-import ItemDetailsPagerVictoryChartPageView from '../&page/victory-chart';
 
 interface ItemDetailsPagerWrapperViewProps {
   item: JoinedItems;
@@ -29,7 +29,7 @@ const ItemDetailsPagerWrapperView = memo(function ItemDetailsPagerWrapperView({
 
     return graphValueFieldArray.map((valueField, index) => (
       <View style={styles.page} key={index + 1} collapsable={false}>
-        <ItemDetailsPagerVictoryChartPageView
+        <ItemDetailsPagerChartPageView
           discountsData={item.discounts!}
           valueField={valueField as 'discount' | 'discountPrice' | 'discountRate'}
         />
