+++ apps/native/package.json
@@ -37,7 +37,7 @@
     "@shopify/react-native-skia": "^1.5.0",
     "@tanstack/react-query": "^5.59.16",
     "date-fns": "^3.6.0",
-    "expo": "52.0.0-preview.7",
+    "expo": "52.0.0-preview.8",
     "expo-apple-authentication": "~7.0.1",
     "expo-build-properties": "~0.13.1",
     "expo-constants": "~17.0.2",
@@ -66,7 +66,6 @@
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "4.0.0-beta.14",
     "react-native-svg": "15.8.0",
-    "react-native-svg-charts": "^5.4.0",
     "react-native-tab-view": "^3.5.2",
     "react-native-unistyles": "^2.11.0",
     "react-native-url-polyfill": "^2.0.0",
